Backport 1.54 1.53 1.52 1.51 1.50 and 1.49:
Try to improve the explanation of the "raise" statement and how its arguments
are interpreted.
This closes SF bug #532467.


Fix Typo.


Reword explanation of global statement since an undeclared global is a
free variable and is subject to those rules.


Note the sole case in which the ban on "from ... import *" within a
function is enforced.


Remove the following restriction:
    Names bound by import statements may not occur in global
    statements in the same scope.
Why not?



Note that it is illegal to delete a cell variable.
Note that deleteing an unbound local will raise a NameError.

--HG--
branch : 2.2
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/release22-maint%4029041