Backport, at the reqest of Kevin Jacobs:
- Changed new-style class instantiation so that when C's __new__
  method returns something that's not a C instance, its __init__ is
  not called.  [SF bug #537450]

XXX This is arguably a semantic change, but it's hard to imagine a
reason for wanting to depend on the old behavior.  If problems with
this are reported within a week of the release of 2.2.2 beta 1, we may
revert this change.

--HG--
branch : 2.2
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/release22-maint%4029122