Fix from Greg Chapman from SF bug #695651: a complex subclass constructor, when passed a single complex argument, returns the argument unchanged.  This should be done only for the complex base class; a complex subclass should of course cast the value to the subclass in this case.
The fix also revealed a segfault in complex_getnewargs(): the argument
for the Py_BuildValue() format code "D" is the *address* of a
Py_complex struct, not the value.  (This corroborated by the API
documentation.)

I expect this needs to be backported to 2.2.3.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4031679