SF bug #855317:  unittest: 5.3.9 Getting Extended Error Information
The example code did not work and could not easily be made to work.
Since the docs were already complex and the feature was not used
(it took two years for the errors to surface), we decided to dedocument
it entirely, leaving unittest cleaner than before.

--HG--
branch : 2.3
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/release23-maint%4034865